[SPARK-46593][PS][TESTS] Refactor `data_type_ops` tests
### What changes were proposed in this pull request?
Refactor `data_type_ops` tests: deduplicate the `OpsTestBase`

### Why are the changes needed?
make `OpsTestBase` reusable and reuse it in the parity tests

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44592 from zhengruifeng/ps_test_data_type_ops.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
