+++ python/pyspark/pandas/tests/connect/data_type_ops/test_parity_boolean_ops.py
@@ -16,19 +16,17 @@
 #
 import unittest
 
-from pyspark import pandas as ps
 from pyspark.pandas.tests.data_type_ops.test_boolean_ops import BooleanOpsTestsMixin
-from pyspark.pandas.tests.connect.data_type_ops.testing_utils import OpsTestBase
-from pyspark.testing.pandasutils import PandasOnSparkTestUtils
+from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
 class BooleanOpsParityTests(
-    BooleanOpsTestsMixin, PandasOnSparkTestUtils, OpsTestBase, ReusedConnectTestCase
+    BooleanOpsTestsMixin,
+    OpsTestBase,
+    ReusedConnectTestCase,
 ):
-    @property
-    def psdf(self):
-        return ps.from_pandas(self.pdf)
+    pass
 
 
 if __name__ == "__main__":
