+++ python/pyspark/pandas/tests/data_type_ops/testing_utils.py
@@ -27,7 +27,7 @@ from pyspark.pandas.typedef.typehints import (
     extension_float_dtypes_available,
     extension_object_dtypes_available,
 )
-from pyspark.testing.pandasutils import ComparisonTestBase
+from pyspark.testing.pandasutils import PandasOnSparkTestCase
 
 if extension_dtypes_available:
     from pandas import Int8Dtype, Int16Dtype, Int32Dtype, Int64Dtype
@@ -39,7 +39,7 @@ if extension_object_dtypes_available:
     from pandas import BooleanDtype, StringDtype
 
 
-class OpsTestBase(ComparisonTestBase):
+class OpsTestBase(PandasOnSparkTestCase):
     """The test base for arithmetic operations of different data types."""
 
     @property
@@ -100,6 +100,10 @@ class OpsTestBase(ComparisonTestBase):
     @property
     def pdf(self):
         return pd.concat([self.numeric_pdf, self.non_numeric_pdf], axis=1)
+
+    @property
+    def psdf(self):
+        return ps.from_pandas(self.pdf)
 
     @property
     def df_cols(self):
